Hide balance at period start on loans
Add condition to check if the balance at period start is not null on a loan. This has been added for when the company is a single year filer this field will not be populated and therefore not displayed. This was causing an issue when rendering the review page.

BI-5145